Added traffic trace parser and did some refactoring

+++ include/packet.h
@@ -7,22 +7,22 @@
 #include <stdlib.h>
 
 /**
-* Standard time to live length 
-*/
+ * Standard time to live length
+ */
 #define TTL_MAX 255
 
 enum packet_state_t { PS_PENDING, PS_IN_NETWORK, PS_TERMINATED };
 
-/** 
-* Network packet object 
-*/
+/**
+ * Network packet object
+ */
 struct packet_t {
   size_t send_id;    // ID of the sending host
   size_t receive_id; // ID of the receiving host
 
-  uint64_t start_time; // start time of packet
-  uint64_t curr_time;  // completion time of packet
-  uint64_t size;       // size of packet in bytes
+  double start_time; // start time of packet
+  double curr_time;  // completion time of packet
+  uint64_t size;     // size of packet in bytes
 
   /* time to live */
   size_t ttl;
@@ -40,13 +40,14 @@ struct packet_t {
 };
 
 /**
-* get the next incoming packet, store its data and evaluate the incoming time (considering all the delays)
-*/
+ * get the next incoming packet, store its data and evaluate the incoming time
+ * (considering all the delays)
+ */
 void packet_traverse_next(struct packet_t *pp);
 
 /**
-* Given the list of the incoming packets, get the next one (earliest) in line
-*/
+ * Given the list of the incoming packets, get the next one (earliest) in line
+ */
 struct packet_t *packet_get_earliest(struct packet_t **const pparr,
                                      const size_t n_pparr);
