+++ include/switch.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 
 struct switch_t {
-  uint64_t last_dequeue_time;
+  double last_dequeue_time;
   double microsec_line_rate;
 };
 
@@ -13,8 +13,8 @@ struct switch_t {
  * get the delay time for the current switch, incorporating the service delay
  * and wait time
  */
-uint64_t switch_get_delay(struct switch_t *s, uint64_t current_time,
-                          uint64_t pkt_size);
+double switch_get_delay(struct switch_t *s, double current_time,
+                        double pkt_size);
 
 /**
  * initialize a switch object, given the line rate in microseconds
